+++ src/main/resources/templates/adminService.html
@@ -12,7 +12,7 @@
 
     <link rel="stylesheet" href="../static/tailwind/font.css">
 
-    <title>Title</title>
+    <title>Service Management</title>
     <style>
         input[type="number"]::-webkit-outer-spin-button,
         input[type="number"]::-webkit-inner-spin-button {
