+++ src/main/resources/templates/fragments/adminHeader.html
@@ -33,7 +33,6 @@
                 class="bx bx-user mr-2"></i>Customer</a>
           </li>
           <li class="relative flex hover:bg-gray-300 cursor-pointer">
-
             <a href=" /see-leave-report" class="block font-bold text-[16px] text-blue-bootstrap pl-2 py-2"><i
                 class="bx bx-calendar-event mr-2" id="see-leave-report"></i>Leave Report</a>
             <div id="notificationDot" class="bg-red-500 rounded-full w-[20px] h-[20px] text-center mr-2 hidden">
@@ -42,28 +41,29 @@
             </div>
           </li>
           <li class="relative">
-
             <a href="/see-all-order-by-admin"
               class="block font-bold text-[16px] text-blue-bootstrap px-2 py-2 hover:bg-gray-300"><i
                 class="bx bx-cart mr-2"></i>Order</a>
           </li>
           <li class="relative">
-
             <a href="/see-all-services"
               class="block font-bold text-[16px] text-blue-bootstrap px-2 py-2 hover:bg-gray-300"><i
                 class="bx bx-book-content mr-2"></i>Services</a>
           </li>
           <li class="relative">
-
             <a href="/see-all-categories"
               class="block font-bold text-[16px] text-blue-bootstrap px-2 py-2 hover:bg-gray-300 rounded-b-lg"><i
                 class="bx bx-category-alt mr-2"></i>Category</a>
           </li>
           <li class="relative">
-
             <a href="/see-all-cancel-request"
-               class="block font-bold text-[16px] text-blue-bootstrap px-2 py-2 hover:bg-gray-300 rounded-b-lg"><i
-                    class="bx bx-category-alt mr-2"></i>Cancel Request</a>
+              class="block font-bold text-[16px] text-blue-bootstrap px-2 py-2 hover:bg-gray-300 rounded-b-lg"><i
+                class="bx bx-category-alt mr-2"></i>Cancel Request</a>
+          </li>
+          <li class="relative">
+            <a href="/dashboard"
+              class="block font-bold text-[16px] text-blue-bootstrap px-2 py-2 hover:bg-gray-300 rounded-b-lg"><i
+                class="bx bx-category-alt mr-2"></i>Dashboard</a>
           </li>
         </ul>
       </div>
